+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 
 **Bug Fixes**:
 
+- Do not drop envelopes for unparsable project configs. ([#3770](https://github.com/getsentry/relay/pull/3770))
+
+## 24.7.0
+
+**Bug Fixes**:
+
 - Fixes raw OS description parsing for iOS and iPadOS originating from the Unity SDK. ([#3780](https://github.com/getsentry/relay/pull/3780))
 - Fixes metrics dropped due to missing project state. ([#3553](https://github.com/getsentry/relay/issues/3553))
 - Incorrect span outcomes when generated from a indexed transaction quota. ([#3793](https://github.com/getsentry/relay/pull/3793))
@@ -24,6 +30,8 @@
 - Remove experimental double-write from spans to transactions. ([#3801](https://github.com/getsentry/relay/pull/3801))
 - Add feature flag to disable replay-video events. ([#3803](https://github.com/getsentry/relay/pull/3803))
 - Write the envelope's Dynamic Sampling Context (DSC) into event payloads for debugging. ([#3811](https://github.com/getsentry/relay/pull/3811))
+- Decrease max allowed segment_id for replays to one hour. ([#3280](https://github.com/getsentry/relay/pull/3280))
+- Extract a duration light metric for spans without a transaction name tag. ([#3772](https://github.com/getsentry/relay/pull/3772))
 
 ## 24.6.0